# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Asset Register API",
        "description": "Laravel 13 / PHP 8.5 hexagonal architecture REST API for managing assets and contracts.",
        "contact": {
            "name": "Asset Register Team"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
        "/api/v1/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "List all assets",
                "operationId": "fcc308d0c17a250345d724573652dc9a",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AssetResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Create an asset",
                "operationId": "fac1f7ac7de64db04b7fe920c6390439",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateAsset"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/assets/{id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get a single asset",
                "operationId": "20cda8496a0323495dffb4e74fce5b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update an asset",
                "operationId": "11348b6ca9a481abe2854ff0060f2caf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateAsset"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Soft-delete an asset",
                "operationId": "ca810826e6a33d32745aa77dad70edce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "409": {
                        "description": "Has active assignments"
                    }
                }
            }
        },
        "/api/v1/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "List all contracts",
                "operationId": "1fb7dc3a9ffbd76dafb271940a493b55",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ContractResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Create a contract",
                "operationId": "d7224ed60662f438f7d9eb6418169c23",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateContract"
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/contracts/{id}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Get contract detail with embedded assets",
                "operationId": "fb7795ed64b1f2e5c2a92587f15b47e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractDetailResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Update a contract",
                "operationId": "dc6cac5201bca92dd435026cbf47d4d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateContract"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Delete a contract (cascades contract_assets, preserves assets)",
                "operationId": "e08c58aa48bb604ee4727b67938645e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/v1/contracts/{id}/assets": {
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Assign an asset to a contract",
                "operationId": "1c832e067033493e8f92b0114f900989",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/AssignAsset"
                },
                "responses": {
                    "201": {
                        "description": "Assigned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContractDetailResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/contracts/{id}/assets/{assetId}": {
            "delete": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Remove an asset from a contract",
                "operationId": "040355bc18e6014f920bb74e8a9813f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "assetId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AssetResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "manufacturer": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ContractDetailResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "contract_number": {
                        "type": "string"
                    },
                    "client_name": {
                        "type": "string"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "assets": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "asset_id": {
                                    "type": "string",
                                    "format": "uuid"
                                },
                                "serial_number": {
                                    "type": "string"
                                },
                                "asset_name": {
                                    "type": "string"
                                },
                                "manufacturer": {
                                    "type": "string"
                                },
                                "model": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ContractResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "contract_number": {
                        "type": "string"
                    },
                    "client_name": {
                        "type": "string"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "AssignAsset": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "asset_id",
                                "serial_number"
                            ],
                            "properties": {
                                "asset_id": {
                                    "type": "string",
                                    "format": "uuid",
                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                },
                                "serial_number": {
                                    "type": "string",
                                    "example": "SN-2026-001"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "CreateAsset": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "manufacturer",
                                "model"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Server Alpha"
                                },
                                "manufacturer": {
                                    "type": "string",
                                    "example": "Dell"
                                },
                                "model": {
                                    "type": "string",
                                    "example": "PowerEdge R750"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "CreateContract": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "contract_number",
                                "client_name",
                                "start_date"
                            ],
                            "properties": {
                                "contract_number": {
                                    "type": "string",
                                    "example": "C-2026-001"
                                },
                                "client_name": {
                                    "type": "string",
                                    "example": "Acme Corp"
                                },
                                "start_date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2026-01-01"
                                },
                                "end_date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2026-12-31",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateAsset": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "manufacturer",
                                "model"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Server Alpha v2"
                                },
                                "manufacturer": {
                                    "type": "string",
                                    "example": "Dell"
                                },
                                "model": {
                                    "type": "string",
                                    "example": "PowerEdge R750"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "UpdateContract": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "contract_number",
                                "client_name",
                                "start_date"
                            ],
                            "properties": {
                                "contract_number": {
                                    "type": "string",
                                    "example": "C-2026-001"
                                },
                                "client_name": {
                                    "type": "string",
                                    "example": "Acme Corp"
                                },
                                "start_date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2026-01-01"
                                },
                                "end_date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2026-12-31",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "tags": [
        {
            "name": "Assets",
            "description": "Asset management"
        },
        {
            "name": "Contracts",
            "description": "Contract management"
        }
    ]
}